fix: changes service to http2 from http (#32)
If Service Protocol is used as HTTP rather than HTTP2, the GRPC will
fail. Since, Platform is a gRPC service, it is best we use HTTP2.

I ran into this by setting up a Istio Ingress Gateway, and the gRPC call
will just end up hanging until the context deadline kills it. Switching
the Service to HTTP2 or GRPC fixed it.