chore(ci): Add workflows to support backporting (#158)
Introduces configuration files for release-please to support different
versioning strategies for the main branch and backport branches. The
main configuration is updated to always bump the minor version, and a
new configuration for the platform package is added for backporting,
which will always bump the patch version. Adds workflows to support
backporting and release branches.